Added remove coins feature Doesn't check if have enough coins, just removes. Checking of coins is handeled else where.
